validacao seguranca

+++ flashcard/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render,redirect
+from django.shortcuts import render,redirect,get_object_or_404
 from .models import Categoria, Flashcard
 from django.http import HttpResponse
 from django.contrib.messages import constants
@@ -48,8 +48,11 @@ def novo_flashcard(request):
     return redirect('/flashcard/novo_flashcard')
   
 def deletar_flashcard(request, id):
-  flashcard = Flashcard.objects.get(id=id)
-  flashcard.delete()
-  messages.add_message(request, constants.SUCCESS, 'Seu FlashCard foi deletado com Sucesso!')
+    flashcard = get_object_or_404(Flashcard, id=id)
+    if flashcard.user != request.user:
+        messages.add_message(request, constants.ERROR, 'Você não tem permissão para excluir este FlashCard.')
+    else:
+        flashcard.delete()
+        messages.add_message(request, constants.SUCCESS, 'Seu FlashCard foi deletado com Sucesso!')
 
-  return redirect('/flashcard/novo_flashcard/')
+    return redirect('/flashcard/novo_flashcard/')
